fold ignore_eos, ebnf, anthropic suites onto the shared openai server
